delete: 중복되는 service, test code 삭제
